consolidated helper functions into single file
